Feature(403006): split into two component
