+++ src/app/components/workflow-diagram/workflow-diagram.component.scss
@@ -8,62 +8,6 @@
     --header-bg-color: #1F4B99; /* Dark blue */
   }
 
-#default-sidebar {
-    background-color: #f4f4f9;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    padding: 20px;
-    height: 100vh;
-    overflow-y: auto;
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  #default-sidebar .title {
-    font-size: 1.5rem;
-    color: #333;
-    margin-bottom: 1rem;
-    text-align: center;
-    width: 100%;
-  }
-  
-  #default-sidebar .content {
-    flex: 1;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  #default-sidebar .center-align {
-    text-align: center;
-  }
-  
-  #default-sidebar .close-btn {
-    background-color: #0f2c60;
-    color: white;
-    border-radius: 6px;
-    padding: 10px 20px;
-    border: none;
-    cursor: pointer;
-    transition: background-color 0.3s;
-  }
-  
-  #default-sidebar .close-btn:hover {
-    background-color: #005a9e;
-  }
-
-  .sidebar-header {
-    font-size: 18px;        /* Example: Set font size */
-    font-weight: bold;      /* Example: Set font weight */
-    padding: 10px;          /* Example: Add some padding */
-    color: #000000;            /* Example: Set text color */
-    text-align: center;     /* Example: Center the text */
-  }
-
   /* Update the styles according to your requirements */
 .boolean-controls {
     display: flex;
@@ -82,21 +26,6 @@
       font-size: 14px;
       box-sizing: border-box;
     }
-  }
-
-  .form-header {
-    background-color: #0f2c60; /* Dark blue background */
-    color: white; /* White text */
-    padding: 10px;
-    text-align: center;
-    font-weight: bold;
-    font-size: 18px;
-  }
-
-  .form-footer {
-    background-color: #0f2c60; /* Same background color as header */
-    padding: 10px;
-    text-align: center;
   }
 
   .full-size-form {
