+++ src/app/components/workflow-diagram/workflow-diagram.component.ts
@@ -1,4 +1,4 @@
-import { AfterViewInit, Component, viewChild, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, EventEmitter, Input, Output, viewChild, ViewChild } from '@angular/core';
 import { ComplexHierarchicalTree, ConnectionPointOrigin, ConnectorConstraints, ConnectorModel, DecoratorModel, Diagram,  DiagramComponent, DiagramModule, HtmlModel, IClickEventArgs, IExportOptions, LayoutModel, LineDistribution, Node, NodeModel, PrintAndExport, SelectorConstraints, SelectorModel, SnapSettingsModel, TextModel, UserHandleEventsArgs, UserHandleModel } from '@syncfusion/ej2-angular-diagrams';
 import { FieldDetails, FieldOptionDetail, FieldValidation, MessageDetails, RuleData, RuleData2 } from '../../models/appModel';
 import { RULE_DATA, RULE_DATA2, RULE_DATA3 } from '../../data/rule-data';
@@ -17,6 +17,8 @@ import { TextFormatEnum, ChatWorkflowEditorTypeEnum, ChatWorkflowBlockTypeEnum }
 import { SwitchModule } from '@syncfusion/ej2-angular-buttons';
 import sampleWorkflowData from '../../data/sample-workflow-data.json'; // Adjust the path as needed
 import { AsyncSettingsModel, FileInfo, Uploader } from '@syncfusion/ej2-inputs';
+import { WorkflowSidebarComponent } from '../workflow-sidebar/workflow-sidebar.component';  // Import child component
+
 
 
 Diagram.Inject(ComplexHierarchicalTree, LineDistribution, PrintAndExport);
@@ -24,7 +26,7 @@ Diagram.Inject(ComplexHierarchicalTree, LineDistribution, PrintAndExport);
 @Component({
   selector: 'app-workflow-diagram',
   standalone: true,
-  imports: [DiagramModule, DialogModule, DropDownButtonModule, CommonModule, ListViewModule, SidebarModule, FormsModule, DropDownListModule, MultiSelectModule, NumericTextBoxModule, TextBoxModule, TextAreaModule, DatePickerModule, DateTimePickerModule, SwitchModule, ToolbarModule, UploaderModule],
+  imports: [DiagramModule, DialogModule, DropDownButtonModule, CommonModule, ListViewModule, DropDownListModule, MultiSelectModule, NumericTextBoxModule, TextBoxModule, TextAreaModule, DatePickerModule, DateTimePickerModule, SwitchModule, ToolbarModule, UploaderModule, WorkflowSidebarComponent],
   templateUrl: './workflow-diagram.component.html',
   styleUrl: './workflow-diagram.component.scss'
 })
@@ -32,19 +34,16 @@ export class WorkflowDiagramComponent implements AfterViewInit{
   @ViewChild('diagram') diagram!: DiagramComponent;
   @ViewChild('dropdownbutton') dropdownbutton!: DropDownButtonComponent;
   @ViewChild('listview') listView!: ListViewComponent;
-  @ViewChild('sidebar') sidebar?: SidebarComponent;
-  @ViewChild('ddlTextFormat') ddlTextFormat!: DropDownListComponent
+  @ViewChild('workflowSidebar') sidebarComponent!: WorkflowSidebarComponent;
 
+  public chatWorkflowEditorTypeEnum = ChatWorkflowEditorTypeEnum; 
+  public chatWorkflowBlockTypeEnum = ChatWorkflowBlockTypeEnum;
   // public data: RuleData[] = RULE_DATA;
   public data: RuleData2[] = RULE_DATA3;
   public nodes: NodeModel[] = [];
   public connectors: ConnectorModel[] = [];
   public closeOnDocumentClick: boolean = true;
   public sidebarInput: string = '';
-  public newNodeWidth: number = 100;
-  public newNodeHeight: number = 100;
-  // public newNodeInfo: RuleData2 = {};
-  public newNode : NodeModel = {}
   public ddlFields: Object = { text: 'label', value: 'value' };
 
   public handles: UserHandleModel[] = [
@@ -63,33 +62,20 @@ export class WorkflowDiagramComponent implements AfterViewInit{
   public headertitle = 'Block';
   public animation: Object  = { duration: 0};
   public isParentListItem : boolean = false;
-  public type: string = 'Push';
-  public width: string ='280px';
-  public sidebarHeader: string = '';
-  public nodeBlockType: number = 0;
-  public nodeEditType: number = 0;
-  public nodeType: string = '';
+  
 
   private nodeIdCounter: number = 0;
   private connectorIdCounter: number = 0;
-  public buttons: Array<{ label: string, value: string, description:string|null }> = [];
-  public options: Array<{ label: string, value: string }> = [];
-  public sideBarLabel: string = '';
-  public sideBarDescription: string = '';
-  public sideBarPlaceholder: string = '';
   public newNodeData: RuleData2[] = [];
-  public fieldOptionMinValue: number = 1;
-  public fieldOptionMaxValue: number = 1;
-  public fieldOptionRegexValue: string = '';
-  public fromDate: Date = new Date();
-  public toDate: Date = new Date();
-  public chatWorkflowEditorTypeEnum = ChatWorkflowEditorTypeEnum; 
-  public chatWorkflowBlockTypeEnum = ChatWorkflowBlockTypeEnum;
   textFormatDDLOptions: Array<{ text: string, value: number }>;
   ddlTextFormatFields: Object = { text: 'text', value: 'value' };
-  public checkedIsPrivate: boolean = false;
-  public customMessage: string = '';
+
   public uploadObject: Uploader;
+
+  public sidebarHeader!: string;
+  public nodeBlockType!: number;
+  public nodeEditType!: number;
+  public clickedNodeId!: string;
 
   // // Async settings for file upload
   // public asyncSettings: AsyncSettingsModel = {
@@ -223,467 +209,28 @@ export class WorkflowDiagramComponent implements AfterViewInit{
       else if(this.diagram.selectedItems.userHandles){
         this.diagram.selectedItems.userHandles[0].visible = false;
       }
-      // const clickedNodeId = args.actualObject.id;
-      // this.addNodeAndConnect(clickedNodeId); // Add and connect a new node
+       this.clickedNodeId = clickedNode.id;
     }
   }
 
+  public getDiagramLength(): number{
+    return this.diagram.nodes.length;
+  }
+
   // Method to add a new node and connect it
-  addNodeAndConnect(sourceNodeId: string): void {
-    switch(this.nodeBlockType){
-      case (this.chatWorkflowBlockTypeEnum.SendTextMessage): {
+  public onaddNodeAndConnect([sourceNodeId, newNode]: [string, NodeModel]): void {
 
-        let messageInfo: MessageDetails = {
-          text: this.customMessage,
-          isPrivate: this.checkedIsPrivate,
-          textFormat: this.ddlTextFormat.value as TextFormatEnum
-        }
-        let newNodeInfo: RuleData2 = {
-          id : this.diagram.nodes.length + 1,
-          chatWorkflowId : 1,
-          successWorkflowId : null,
-          successRuleId : null,
-          isActive : true,
-          chatWorkflowBlockId : 6,
-          chatWorkflowEditorTypeId : null,
-          fieldDetails : null,
-          branchDetails: null,
-          messageDetails: messageInfo,
-          fieldOptionDetails: null
-        };
-
-        this.newNodeHeight= 100; // Set a default height for the new node
-        this.newNodeWidth = 200;
-        this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-        break;
-      } 
-      case (this.chatWorkflowBlockTypeEnum.GetPickerInput):{
-        switch(this.nodeEditType){
-          case (this.chatWorkflowEditorTypeEnum.Boolean): {
-            let fieldInfo: FieldDetails = {
-              description : this.sideBarDescription ,
-              label : this.sideBarLabel,
-              placeholder : this.sideBarPlaceholder
-            };
-            let newNodeInfo: RuleData2 = {
-              id : this.diagram.nodes.length + 1,
-              chatWorkflowId : 1,
-              successWorkflowId : 1,
-              successRuleId : null,
-              isActive : true,
-              chatWorkflowBlockId : 4,
-              chatWorkflowEditorTypeId : 1,
-              fieldDetails : fieldInfo,
-              branchDetails: null,
-              messageDetails: null,
-              fieldOptionDetails: null
-            };
-    
-            this.newNodeHeight= 100; // Set a default height for the new node
-            this.newNodeWidth = 150;
-            this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-            break;
-          }
-          case(this.chatWorkflowEditorTypeEnum.Buttons): {
-            let fieldInfo: FieldDetails = {
-              description : this.sideBarDescription ,
-              label : this.sideBarLabel,
-              placeholder : this.sideBarPlaceholder
-            };
-    
-            // Mapping buttons to fieldOptionInfo
-            let fieldOptionInfo: FieldOptionDetail[] = this.buttons.map(button => {
-              return {
-                label: button.label,
-                value: button.value
-              };
-            });
-    
-            let newNodeInfo: RuleData2 = {
-              id : this.diagram.nodes.length + 1,
-              chatWorkflowId : 1,
-              successWorkflowId : 1,
-              successRuleId : null,
-              isActive : true,
-              chatWorkflowBlockId : 4,
-              chatWorkflowEditorTypeId : 2,
-              fieldDetails : fieldInfo,
-              branchDetails: null,
-              messageDetails: null,
-              fieldOptionDetails: fieldOptionInfo
-            };
-    
-            this.newNodeHeight= 100 + (this.buttons.length * 20); // Set a default height for the new node
-            this.newNodeWidth = 200;
-            this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-            break;
-          }
-          case this.chatWorkflowEditorTypeEnum.DropDown: {
-              let fieldInfo: FieldDetails = {
-                description : this.sideBarDescription ,
-                label : this.sideBarLabel,
-                placeholder : this.sideBarPlaceholder
-              };
-        
-              // Mapping buttons to fieldOptionInfo
-              let fieldOptionInfo: FieldOptionDetail[] = this.buttons.map(button => {
-                return {
-                  label: button.label,
-                  value: button.value
-                };
-              });
-        
-              let newNodeInfo: RuleData2 = {
-                id : this.diagram.nodes.length + 1,
-                chatWorkflowId : 1,
-                successWorkflowId : 1,
-                successRuleId : null,
-                isActive : true,
-                chatWorkflowBlockId : 4,
-                chatWorkflowEditorTypeId : 3,
-                fieldDetails : fieldInfo,
-                branchDetails: null,
-                messageDetails: null,
-                fieldOptionDetails: fieldOptionInfo
-              };
-        
-              this.newNodeHeight= 100; // Set a default height for the new node
-              this.newNodeWidth = 200;
-              this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-              break;
-          }
-          case (this.chatWorkflowEditorTypeEnum.MultiSelect): {
-              let fieldValidationInfo: FieldValidation = {
-                min: this.fieldOptionMinValue.toString(),
-                max: this.fieldOptionMaxValue.toString()
-              };
-              let fieldInfo: FieldDetails = {
-                description : this.sideBarDescription ,
-                label : this.sideBarLabel,
-                placeholder : this.sideBarPlaceholder,
-                fieldValidation: fieldValidationInfo
-              };
-    
-              // Mapping buttons to fieldOptionInfo
-              let fieldOptionInfo: FieldOptionDetail[] = this.buttons.map(button => {
-                return {
-                  label: button.label,
-                  value: button.value
-                };
-              });
-    
-              let newNodeInfo: RuleData2 = {
-                id : this.diagram.nodes.length + 1,
-                chatWorkflowId : 1,
-                successWorkflowId : 1,
-                successRuleId : null,
-                isActive : true,
-                chatWorkflowBlockId : 4,
-                chatWorkflowEditorTypeId : 4,
-                fieldDetails : fieldInfo,
-                branchDetails: null,
-                messageDetails: null,
-                fieldOptionDetails: fieldOptionInfo
-              };
-    
-              this.newNodeHeight= 100; // Set a default height for the new node
-              this.newNodeWidth = 200;
-              this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-              break;
-          } 
-          case (this.chatWorkflowEditorTypeEnum.List): {
-              let fieldInfo: FieldDetails = {
-                description : this.sideBarDescription ,
-                label : this.sideBarLabel,
-                placeholder : this.sideBarPlaceholder,
-              };
-    
-              // Mapping buttons to fieldOptionInfo
-              let fieldOptionInfo: FieldOptionDetail[] = this.buttons.map(button => {
-                return {
-                  label: button.label,
-                  value: button.value,
-                  description: button.description !== null ? button.description : null
-                };
-              });
-    
-              let newNodeInfo: RuleData2 = {
-                id : this.diagram.nodes.length + 1,
-                chatWorkflowId : 1,
-                successWorkflowId : 1,
-                successRuleId : null,
-                isActive : true,
-                chatWorkflowBlockId : 4,
-                chatWorkflowEditorTypeId : 5,
-                fieldDetails : fieldInfo,
-                branchDetails: null,
-                messageDetails: null,
-                fieldOptionDetails: fieldOptionInfo
-              };
-    
-              this.newNodeHeight= 100 + (this.buttons.length * 25); // Set a default height for the new node
-              this.newNodeWidth = 200;
-              this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-              break;
-          } 
-        }
-        break;
-      }
-      case (this.chatWorkflowBlockTypeEnum.GetTextInput):{
-        switch(this.nodeEditType){
-          case (this.chatWorkflowEditorTypeEnum.Text): {
-            let fieldValidationInfo: FieldValidation = {
-              max: this.fieldOptionMaxValue.toString()
-            };
-            let fieldInfo: FieldDetails = {
-              description : this.sideBarDescription ,
-              label : this.sideBarLabel,
-              placeholder : this.sideBarPlaceholder,
-              fieldValidation: fieldValidationInfo
-            };
-  
-            let newNodeInfo: RuleData2 = {
-              id : this.diagram.nodes.length + 1,
-              chatWorkflowId : 1,
-              successWorkflowId : 1,
-              successRuleId : null,
-              isActive : true,
-              chatWorkflowBlockId : 5,
-              chatWorkflowEditorTypeId : 7,
-              fieldDetails : fieldInfo,
-              branchDetails: null,
-              messageDetails: null,
-              fieldOptionDetails: null
-            };
-  
-            this.newNodeHeight= 100; // Set a default height for the new node
-            this.newNodeWidth = 200;
-            this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-            break;
-          } 
-          case (this.chatWorkflowEditorTypeEnum.TextArea): {
-              let fieldValidationInfo: FieldValidation = {
-                max: this.fieldOptionMaxValue.toString()
-              };
-              let fieldInfo: FieldDetails = {
-                description : this.sideBarDescription ,
-                label : this.sideBarLabel,
-                placeholder : this.sideBarPlaceholder,
-                fieldValidation: fieldValidationInfo
-              };
-    
-              let newNodeInfo: RuleData2 = {
-                id : this.diagram.nodes.length + 1,
-                chatWorkflowId : 1,
-                successWorkflowId : 1,
-                successRuleId : null,
-                isActive : true,
-                chatWorkflowBlockId : 5,
-                chatWorkflowEditorTypeId : 8,
-                fieldDetails : fieldInfo,
-                branchDetails: null,
-                messageDetails: null,
-                fieldOptionDetails: null
-              };
-    
-              this.newNodeHeight= 100; // Set a default height for the new node
-              this.newNodeWidth = 200;
-              this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-              break;
-          } 
-          case (this.chatWorkflowEditorTypeEnum.Date): {
-            const today = new Date();
-            // Format today's date as YYYY-MM-DD
-            const minDate = today.toISOString().split('T')[0];
-  
-            // Calculate max date, e.g., 30 days from today
-            const maxDateObj = new Date(today);
-            maxDateObj.setDate(today.getDate() + 30);
-            const maxDate = maxDateObj.toISOString().split('T')[0];
-  
-            let fieldValidationInfo: FieldValidation = {
-              // min: this.fromDate.toString(),
-              // max: this.toDate.toString()
-              
-              min: minDate,
-              max: maxDate
-            };
-            let fieldInfo: FieldDetails = {
-              description : this.sideBarDescription ,
-              label : this.sideBarLabel,
-              placeholder : this.sideBarPlaceholder,
-              fieldValidation: fieldValidationInfo
-            };
-  
-            let newNodeInfo: RuleData2 = {
-              id : this.diagram.nodes.length + 1,
-              chatWorkflowId : 1,
-              successWorkflowId : 1,
-              successRuleId : null,
-              isActive : true,
-              chatWorkflowBlockId : 5,
-              chatWorkflowEditorTypeId : 9,
-              fieldDetails : fieldInfo,
-              branchDetails: null,
-              messageDetails: null,
-              fieldOptionDetails: null
-            };
-  
-            this.newNodeHeight= 100; // Set a default height for the new node
-            this.newNodeWidth = 200;
-            this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-            break;
-          } 
-          case (this.chatWorkflowEditorTypeEnum.DateTime): {
-              const now = new Date();
-              // Current datetime formatted as ISO string
-              const minDateTime = now.toISOString();
-            
-              // Max datetime, e.g., 30 days from now
-              const maxDateTimeObj = new Date(now);
-              maxDateTimeObj.setDate(now.getDate() + 30);
-              const maxDateTime = maxDateTimeObj.toISOString();
-    
-              let fieldValidationInfo: FieldValidation = {
-                min: minDateTime,
-                max: maxDateTime
-              };
-              let fieldInfo: FieldDetails = {
-                description : this.sideBarDescription ,
-                label : this.sideBarLabel,
-                placeholder : this.sideBarPlaceholder,
-                fieldValidation: fieldValidationInfo
-              };
-    
-              let newNodeInfo: RuleData2 = {
-                id : this.diagram.nodes.length + 1,
-                chatWorkflowId : 1,
-                successWorkflowId : 1,
-                successRuleId : null,
-                isActive : true,
-                chatWorkflowBlockId : 5,
-                chatWorkflowEditorTypeId : 10,
-                fieldDetails : fieldInfo,
-                branchDetails: null,
-                messageDetails: null,
-                fieldOptionDetails: null
-              };
-    
-              this.newNodeHeight= 100; // Set a default height for the new node
-              this.newNodeWidth = 200;
-              this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-              break;
-          }
-          case (this.chatWorkflowEditorTypeEnum.Number): {
-            let fieldValidationInfo: FieldValidation = {
-              min: this.fieldOptionMinValue.toString(),
-              max: this.fieldOptionMaxValue.toString()
-            };
-            let fieldInfo: FieldDetails = {
-              description : this.sideBarDescription ,
-              label : this.sideBarLabel,
-              placeholder : this.sideBarPlaceholder,
-              fieldValidation: fieldValidationInfo
-            };
-  
-            let newNodeInfo: RuleData2 = {
-              id : this.diagram.nodes.length + 1,
-              chatWorkflowId : 1,
-              successWorkflowId : 1,
-              successRuleId : null,
-              isActive : true,
-              chatWorkflowBlockId : 5,
-              chatWorkflowEditorTypeId : 11,
-              fieldDetails : fieldInfo,
-              branchDetails: null,
-              messageDetails: null,
-              fieldOptionDetails: null
-            };
-  
-            this.newNodeHeight= 100; // Set a default height for the new node
-            this.newNodeWidth = 200;
-            this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-            break;
-          } 
-          case (this.chatWorkflowEditorTypeEnum.Decimal): {
-              let fieldValidationInfo: FieldValidation = {
-                min: this.fieldOptionMinValue.toString(),
-                max: this.fieldOptionMaxValue.toString()
-              };
-              let fieldInfo: FieldDetails = {
-                description : this.sideBarDescription ,
-                label : this.sideBarLabel,
-                placeholder : this.sideBarPlaceholder,
-                fieldValidation: fieldValidationInfo
-              };
-    
-              let newNodeInfo: RuleData2 = {
-                id : this.diagram.nodes.length + 1,
-                chatWorkflowId : 1,
-                successWorkflowId : 1,
-                successRuleId : null,
-                isActive : true,
-                chatWorkflowBlockId : 5,
-                chatWorkflowEditorTypeId : 12,
-                fieldDetails : fieldInfo,
-                branchDetails: null,
-                messageDetails: null,
-                fieldOptionDetails: null
-              };
-    
-              this.newNodeHeight= 100; // Set a default height for the new node
-              this.newNodeWidth = 200;
-              this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-              break;
-          } 
-          case (this.chatWorkflowEditorTypeEnum.Regex): {
-              let fieldValidationInfo: FieldValidation = {
-                regex: this.fieldOptionRegexValue
-              };
-              let fieldInfo: FieldDetails = {
-                description : this.sideBarDescription ,
-                label : this.sideBarLabel,
-                placeholder : this.sideBarPlaceholder,
-                fieldValidation: fieldValidationInfo
-              };
-    
-              let newNodeInfo: RuleData2 = {
-                id : this.diagram.nodes.length + 1,
-                chatWorkflowId : 1,
-                successWorkflowId : 1,
-                successRuleId : null,
-                isActive : true,
-                chatWorkflowBlockId : 5,
-                chatWorkflowEditorTypeId : 13,
-                fieldDetails : fieldInfo,
-                branchDetails: null,
-                messageDetails: null,
-                fieldOptionDetails: null
-              };
-    
-              this.newNodeHeight= 100; // Set a default height for the new node
-              this.newNodeWidth = 200;
-              this.newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, newNodeInfo);
-              break;
-          } 
-        }
-        break;
-      }
-    }
-
-    console.log('newNode: '+ this.newNode.addInfo);
-
+    this.nodes.push(newNode);
     // const newNode = this.createNode(this.newNodeWidth, this.newNodeHeight, htmlContent, newNodeInfo); // Width and height as parameters
-
     // Add the new node to the diagram
-    this.diagram.addNode(this.newNode);
+    this.diagram.addNode(newNode);
 
     // Create a new connector to link the new node to the source node
     const newConnectorId = `connector${++this.connectorIdCounter}`;
     const newConnector: ConnectorModel = {
       id: newConnectorId,
       sourceID: sourceNodeId,
-      targetID: this.newNode.id,
+      targetID: newNode.id,
       type: 'Orthogonal',
       style: { strokeColor: '#6BA5D7', strokeWidth: 1 }
     };
@@ -692,31 +239,13 @@ export class WorkflowDiagramComponent implements AfterViewInit{
     this.diagram.addConnector(newConnector);
   }
 
-  createNode(width: number, height: number, nodeInfo: RuleData2): NodeModel {
-    return {
-      id: `node${nodeInfo.id}`,
-      annotations: [],
-      offsetX: 100,
-      offsetY: 70,
-      style: { fill: '#FFFFFF', strokeColor: '#0f2c60', strokeWidth: 5 },
-      height: height,
-      width: width,
-      borderColor: '#0f2c60',
-      borderWidth: 3,
-      shape: { 
-        type: 'HTML',
-      },
-      addInfo: nodeInfo
-    };
-  }
-
   public onUserHandleMouseDown(event: UserHandleEventsArgs) {
     if(event.element.name === 'plusIcon') {
       this.dropdownbutton.toggle();
     }
   }
 
-  onOpenDropDownButton(args: OpenCloseMenuEventArgs) {
+  public onOpenDropDownButton(args: OpenCloseMenuEventArgs) {
     let dropDownContainer = document.querySelector('.dropDown-container') as HTMLElement;
     args.element.parentElement!.style.top = dropDownContainer.getBoundingClientRect().top + dropDownContainer.offsetHeight +'px';
 
@@ -724,14 +253,14 @@ export class WorkflowDiagramComponent implements AfterViewInit{
     args.element.parentElement!.style.left = dropDownContainer.getBoundingClientRect().left - (ulElement.getBoundingClientRect().width / 2)+ (dropDownContainer.getBoundingClientRect().width / 2)+'px' ;
   }
 
-  onBeforeCloseDropDownButton(args: BeforeOpenCloseMenuEventArgs) {
+  public onBeforeCloseDropDownButton(args: BeforeOpenCloseMenuEventArgs) {
     args.cancel = this.isParentListItem;
     if (!(this.isParentListItem || ((args.event.target as HTMLElement).closest(".bc-block-option.e-listview")==null))) {
-      this.sidebar?.show();
+      this.sidebarComponent?.sidebar?.show();
     }
   }
 
-  onSelectListView(args: SelectEventArgs) {
+  public onSelectListView(args: SelectEventArgs) {
     this.isParentListItem = args.item.classList.contains("e-has-child");
     const selectedItem = args.item;
     const selectedItemId = selectedItem.getAttribute('data-uid');
@@ -747,7 +276,7 @@ export class WorkflowDiagramComponent implements AfterViewInit{
     }
   }
 
-  onBeforeOpenDropDownButton() {
+  public onBeforeOpenDropDownButton() {
     this.isParentListItem = false; // The value is reset here, to handle document click case of dropdown
     // Reset ListView to its initial state before opening
     if (this.listView) {
@@ -756,38 +285,7 @@ export class WorkflowDiagramComponent implements AfterViewInit{
     }
   }
 
-  onCloseSideBarClick(): void {
-    this.sidebar?.hide();
-    this.addNodeAndConnect('node1');
-    this.buttons = [];
-  }
-
-  public onSideBarCreated(args: any) {
-    (this.sidebar as SidebarComponent).element.style.visibility = '';
-    (this.sidebar as SidebarComponent).position = "Right";
-  }
-
-  onFormSubmit(): void {
-    // this.addNodeAndConnect('1');
-    this.sidebar?.hide();
-  }
-
-  addButton(label: string, value: string, description: string | null, labelInput: HTMLInputElement, valueInput: HTMLInputElement, descriptionInput: HTMLInputElement | null): void {
-     if (label.trim() && value.trim()) {
-      this.buttons.push({ label, value, description });
-      labelInput.value = '';
-      valueInput.value = '';
-      if (descriptionInput) {
-        descriptionInput.value = '';
-      }
-    }
-  }
-
-  removeButton(index: number): void {
-    this.buttons.splice(index, 1);
-  }
-
-  onClicked(args: ClickEventArgs) {
+  public onClicked(args: ClickEventArgs) {
     if (args.item.text === 'Save') {
       this.download(this.diagram.saveDiagram());
     }
